Compilation on darwin warning/error fix

+++ library/log.c
@@ -45,7 +45,7 @@ void eblob_log_raw_formatted(void *priv, uint32_t mask, const char *msg)
 	localtime_r((time_t *)&tv.tv_sec, &tm);
 	strftime(str, sizeof(str), "%F %R:%S", &tm);
 
-	fprintf(stream, "%s.%06lu %1x: %s", str, tv.tv_usec, mask, msg);
+	fprintf(stream, "%s.%06lu %1x: %s", str, (unsigned long)tv.tv_usec, mask, msg);
 	fflush(stream);
 }
